+++ src/extensions/rules/rulesext.cpp
@@ -93,7 +93,11 @@ RulesClassExtension::RulesClassExtension(const RulesClass *this_ptr) :
     VoxelLightAzimuth(0),
     VoxelLightElevation(DEG_TO_RAD(45)),
     VoxelShadowOffset(6),
-    IsTiberiumStorage(true)
+    IsTiberiumStorage(true),
+    UpgradeVeteranSound(VOC_NONE),
+    UpgradeEliteSound(VOC_NONE),
+    VoxUnitPromoted(VOX_NONE),
+    EliteFlashTimer(0)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("RulesClassExtension::RulesClassExtension - 0x%08X\n", (uintptr_t)(ThisPtr));
 
@@ -686,6 +690,11 @@ bool RulesClassExtension::AudioVisual(CCINIClass &ini)
     VoxelShadowOffset = ini.Get_Float(AUDIOVISUAL, "VoxelShadowOffset", VoxelShadowOffset);
 
     Set_Voxel_Light_Angle(VoxelLightAzimuth, VoxelLightElevation, VoxelShadowOffset);
+
+    UpgradeVeteranSound = ini.Get_VocType(AUDIOVISUAL, "UpgradeVeteranSound", UpgradeVeteranSound);
+    UpgradeEliteSound = ini.Get_VocType(AUDIOVISUAL, "UpgradeEliteSound", UpgradeEliteSound);
+    VoxUnitPromoted = ini.Get_VoxType(AUDIOVISUAL, "VoxUnitPromoted", VoxUnitPromoted);
+    EliteFlashTimer = ini.Get_Int(AUDIOVISUAL, "EliteFlashTimer", EliteFlashTimer);
 
     return true;
 }