+++ src/extensions/techno/technoext.cpp
@@ -71,7 +71,8 @@ TechnoClassExtension::TechnoClassExtension(const TechnoClass *this_ptr) :
     HasOpportunityFireTarget(false),
     LastTargetFrame(Frame),
     IsToResetBurst(false),
-    BurstResetTimer()
+    BurstResetTimer(),
+    LastVeterancy(RANK_NONE)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("TechnoClassExtension::TechnoClassExtension - Name: %s (0x%08X)\n", Name(), (uintptr_t)(This()));